gbm, mesa: Unroll dep lists. #275
libXdmcp actually isn't a dependency of Mesa, it's mentioned in the
libxcb xcb.pc file, in requires private. Since it's a Recommended
dependency of libxcb, and is required for the pkgconf file, it will only
cause issues to skip libXdmcp. Thus, it should be required for libxcb.

For GBM, it's a sad case as it requires xcb, but only realistically
needs libdrm and the Python packages. We'll have to settle for this.