*: Remove <application> tags.
The application tag seemed to just put an extra space on either side of
the word(s) the tags surrounded and seemed to serve no purpose beyond
that. If it served the purpose of jhalfs, well this book is not
supported by jhalfs. The result of the tags just added space in both the
rendered output and in the XML files.

Now, when referring to a package name, we just say the name and don't
surround it via a tag. The only exception is using <command>,
<filename>, <xref>, or <ulink>.