GBM: Move header installations to NVIDIA.
The plan is to add libgbm separate from Mesa in the near future to dodge
several dependencies. It will not include dri_interface.h nor
mesaeglext.h. Since header installation would then look the same, it
makes more sense to offload the headers to NVIDIA, and leave it to Mesa
to install them for a full installation if desired.