shareddeps: Split Graphics Drivers into multiple chapters.
8.  Graphics APIs
9.  Graphics Drivers, Preliminaries
10. Graphics Drivers, NVIDIA
11. Graphics Drivers, Mesa

This was done to set up the groundwork for including multiple NVIDIA
versions to support more NVIDIA GPUs. r590 killed off support for
pre-turing GPUs, meaning anything before GTX 16xx is effectively dead
unless Mesa's nouveau is used. This is a problem because pre-turing GPUs
are everywhere and are still recommended as beginner GPUs. Thus, there's
cause to have older NVIDIA versions, as well as new ones.

The main plan is to split up the NVIDIA chapter somehow. This can be
done via dummies. There are a few issues to bear:
1. To create separate pages for different NVIDIA versions will create a
   lot of duplicate work. This can be resolved via entities.
2. Including multiple NVIDIA versions in a single page will result in
   the NVIDIA page bolstering in size. Already, the NVIDIA page in the
   book is the longest of them all for no good reason beyond the
   complexity of installation. This is a result of our manual
   kernel installation approach. The driver library approach has been
   made much shorter already.
3. CUDA follows NVIDIA in terms of GPU and driver support. Meaning new
   CUDA versions will require an at-least NVIDIA version, often that of
   a beta version. This causes issues since newer CUDA versions can't be
   used unless the driver is upgraded to the minimum. Thus, multiple
   CUDA versions must be provided.
4. Because of issue 3, this causes issues for SLFS and conciseness. As
   such, the plan is to move CUDA to this book. With everything,
   separate pages for separate versions will be the best approach; an
   ugly one, but our best shot.

Entities will be our best friend here.