libaom: Use an alternative method to prevent static libraries.
This follows what we are doing in BLFS now, though in that case it was
required because we were removing the static library and CMake was still
referencing it in files that it installed.

It caused issues over there with libheif, and even though that package
isn't in GLFS, let's fix it over here as well for consistency.

Fixes #354